+++ app/src/main/java/cz/jaro/dpmcb/ui/oblibene/OblibeneState.kt
@@ -1,0 +1,8 @@
+package cz.jaro.dpmcb.ui.oblibene
+
+data class OblibeneState(
+    val nacitaSe: Boolean,
+    val nejake: Boolean,
+    val dnes: List<KartickaState>,
+    val jindy: List<KartickaState>,
+)
