+++ app/src/main/java/cz/jaro/dpmcb/ui/zjr/JizdniRadyState.kt
@@ -1,0 +1,8 @@
+package cz.jaro.dpmcb.ui.zjr
+
+import cz.jaro.dpmcb.data.realtions.OdjezdNizkopodlaznostSpojId
+
+sealed interface JizdniRadyState {
+    object Loading : JizdniRadyState
+    data class Success(val data: List<OdjezdNizkopodlaznostSpojId>) : JizdniRadyState
+}
